Fixed massive bug introduced by new caching system.
Since the old caching system associated with nodes rather than
addresses, it was never possible to encounter cached pages generated
by other parts of the hierarchy.  With the new cached content being
associated with _addresses_, this became possible.  Since any nodes
referenced by that content already had parents pointing to that other
part of the hierarchy, this led to confusing behaviour when navigating
to parent nodes - these would be the "old" parents, not the new ones!

To address this, node parents are now set every time they are
visited (unless they are visited from below).