Add Gemini search link to elpher homepage
This commit adds a "homepage" link to GUS in a similar fashion to
Veronica. It's not actually tested, because I'm not very well-versed
in hacking Emacs packages, and doing something like over-writing an
installed version from Melpa.

Thank you for the great client, and please LMK if you have any
feedback or guidance on how to improve this implementation!