Fill headers
Filling headers is tricky because fill-column is measured in
characters but the font-size for headers is much larger than for
regular text. This change tries to accomodate this by taking the face
for the heading, getting it's font – a string like
"-BE5N-Iosevka-normal-normal-normal-*-29-*-*-*-d-0-iso10646-1" – turns
it into a spec and gets the size from there: 29. We can't use
something like (face-attribute face :height) directly because it might
return a font size in 1/10 point (if an interger) or as a scaling
factor (if a float). Thus, in order to avoid tracking down the face
inheritance tree, we're looking at the font name (a string), in either
XLFD, Fontconfig, or GTK+ format.