Replace Elpher bookmarks with Emacs bookmarks
The keybindings for x and X disappear. M-x elpher-import-bookmarks
imports the old bookmarks into regular Emacs bookmarks. This deletes
all the old code related to bookmarks! This is a big change.

Update the manual to reflect this.

Add Makefile rules to make HTML and PDF files. Fix the texinfo file.
For the PDF output, we cannot use @kbd inside @keycmd, so the entry
for elpher-prev-link cannot use @keycmd, in order to typeset Shift-TAB
correctly.