Change browse-url advice for older releases
package does not offer browse-url-default-handlers, but if you set
The code used to advise browse-url-default-browser if your browse-url
browse-url-browser-function to something else, like eww-browse-url,
then that advice did not apply. The code now advises the function set
in the browse-url-browser-function variable, meaning that it should
advise the right function.

Also added a small doc string to the lambda used as advice.