Prevents indenting gemini lines prefixed only with spaces.
Addresses this comment from the elpher feedback page:
Admittedly, this is a small complaint, but elpher has problems
with indentations in paragraphs in gemini (haven't tested with
gopher). It is supposed to stop indenting after the first line
break, but doesn't do so, instead keeping the indentation for the
whole paragraph. I uploaded a picture showing on the top the
expected result and on the bottom what elpher shows.
https://ttm.sh/if_.png